requirements.txt
Django==1.5.1
South==0.8.1

requests==1.2.0
djangorestframework==2.3.10

nose==1.3.0
django-nose==1.1
factory_boy==1.3.0

ipython==0.13.2
ipdb==0.7

